Test seperating server file into app and server files
